Thu, May 30, 2024, 4:25 PM +05:30

+++ calculator.py
@@ -1,5 +1,5 @@
-x = int(input("x: "))
-y = int(input("y: "))
+x = float(input("x: "))
+y = float(input("y: "))
 
 z = x + y
